pyproject.toml
[project]
name = "astro-lab"
version = "0.1.0"
description = "Astronomical Machine Learning Laboratory with MLflow tracking"
readme = "README.md"
authors = [
    { name = "Bjoern Bethge", email = "bjoern.bethge@gmail.com" }
]
maintainers = [
    { name = "Bjoern Bethge", email = "bjoern.bethge@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "h5py>=3.8.0", # check astropy.io API
    "pydantic>=2.11.6",
    "pyyaml>=6.0.2",
    "markupsafe>=3.0.2",
    # astro
    "astroml>=1.0.2.post1",
    "astrophot>=0.16.13",
    "astropy>=7.1.0",
    "astroquery>=0.4.10",
    "poliastro>=0.7.0",
    "sdss-access>=3.0.8",
    "sgp4>=2.24",
    "numpy>=1.24,<2.0",
    # ml
    "torch>=2.7.1",
    "torch-geometric>=2.6.1",
    "torchinfo>=1.8.0",
    "torchmetrics==1.6.3",
    "marimo-flow>=0.1.1",
    "lightning==2.5.1.post0",
    "lightning-utilities==0.14.3",
    "optuna-integration[pytorch-lightning]>=4.4.0",
    "optuna>=4.4.0",
    # viz
    "bpy>=4.4.0",
    "cosmograph>=0.0.47",
    "pyvista>=0.45.2",
    "open3d>=0.19.0",
    "plotly>=6.1.2",
    "anywidget>=0.9.18",
    "traitlets>=5.14.3",
    "tensordict>=0.8.3",
    "requests>=2.32.4",
    "mathutils>=3.3.0",
    "photutils>=2.2.0",
    "triton-windows<3.4",
    "torchvision>=0.22.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
link-mode = "copy"

# Development dependency groups (PEP 735)
[dependency-groups]
# Core development tools
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "psutil>=6.1.0",
    "pytest-xdist>=3.7.0",
    "pytest-timeout>=2.0.0", # Test timeouts
    "pytest-rerunfailures>=12.0", # Retry failed tests
    "psutil>=6.1.0",
    "gputil>=1.4.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.20.0",
    "griffe-pydantic>=1.1.4",
    "pymdown-extensions>=10.15",
    "mkdocs-mermaid2-plugin>=1.2.1",
    "graphviz>=0.21",
    "pydot>=4.0.1",
]


# PyTorch with CUDA support
[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cuda" }]
torchvision = [{ index = "pytorch-cuda" }]

# Console scripts for CLI tools
[project.scripts]
astro-lab = "astro_lab.cli:main"
astro-test = "scripts.run_tests:main"
